[depends] Add patch removing "png" from supported extensions
The "png" format is for ROMs steganographically encoded as PNGs. However,
when a valid, normal PNG is provided, an assertion fails and Kodi aborts.

The core should be fixed to not assert and fail gracefully, so until it
does, we need to remove steganographic PNG support to not crash Kodi.

Assertion happens here:

  https://github.com/libretro/retro8/blob/master/src/io/stegano.cpp#L265